+++ Example/iOS Example/ObjectiveCCompatibilityChecker.h
@@ -1,0 +1,7 @@
+// We use this file to check if all calls are usable from Objective-C code.
+
+#import <Foundation/Foundation.h>
+
+@interface ObjectiveCCompatibilityCheker : NSObject
+
+@end